% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiedown_pdf_document.R
\name{mmmm}
\alias{mmmm}
\title{Customized RMarkdown Document}
\usage{
mmmm(...)
}
\description{
\code{mmmm()} is the main function of the mmmm package. Use as
\code{output: mmmm::mmmm} in the RMarkdown YAML header.
}
