% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_page_title.R
\name{cd_page_title}
\alias{cd_page_title}
\title{Corporate Design Element for Title Page}
\usage{
cd_page_title(
  title = default(rmarkdown::metadata$title, "Title"),
  subtitle = default(rmarkdown::metadata$subtitle, "Subtitle"),
  date = default(rmarkdown::metadata$date, cd_format_date(Sys.Date()))
)
}
\arguments{
\item{title}{Document title.}

\item{subtitle}{Document subtitle.}

\item{date}{Document creation date.}
}
\description{
Example function to generate a title page
}
