% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision.R
\name{decision}
\alias{decision}
\title{Decision}
\usage{
decision(n1, n2, d1, d2)
}
\arguments{
\item{n1, }{n2 Booleans. n1: (w$children\verb{[[1]]}$n > Nmin)}

\item{d1, }{d2 Numerics. d1: w$children\verb{[[1]]}$dom}
}
\value{
n1\*d1 + n2\*d2 + n1\*d1\*n2\*d2 || n1\*d1 == n2\*d2
}
\description{
Wrapper function to decide wether or not a node's children should be pruned cleanly.
A value of TRUE means the node's children will be pruned. Internal use only.
}
\keyword{internal}
