% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.transitions.R
\name{expand.transitions}
\alias{expand.transitions}
\title{Extracts transitions from a skeleton and expands them to order k.}
\usage{
expand.transitions(skel, A, sep)
}
\arguments{
\item{skel}{A skeleton tree, such as the result of \link{sculptskeleton}.}

\item{A}{Alphabet of possible symbols.}

\item{sep}{Character separator for the skeleton contexts.}
}
\value{
A list, where each element corresponds to the transition possibilities from past w (name) to each symbol u in A.
}
\description{
Extracts transitions from a skeleton object and expands contexts of order less than k to all possible pasts of order k. Internal use.
}
\keyword{internal}
