% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandTransitions.R
\name{expandTransitions}
\alias{expandTransitions}
\title{Extracts transitions from a skeleton and expands them to order k.}
\usage{
expandTransitions(skeleton, A, sep)
}
\arguments{
\item{skeleton}{A result of \link{sculptskeleton}.}

\item{A}{Alphabet.}

\item{sep}{Character separator for the skeleton contexts.}
}
\value{
A list, where each element corresponds to the transition possibilities from past w (name) to each symbol u in A.
}
\description{
Extracts transitions from a skeleton object and expands contexts of order less than k to all possible pasts of order k.
}
