% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_skeleton.R
\name{generate_skeleton}
\alias{generate_skeleton}
\title{Generate skeleton tree}
\usage{
generate_skeleton(
  sample,
  Nmin = 59,
  sensibility,
  alpha,
  asDataFrameTree = T,
  cleantree = T,
  silent = F
)
}
\arguments{
\item{sample}{Data. Either a string or a numeric vector. Should be longer than Nmin and only contain 0 and 1 symbols.}

\item{Nmin}{A transition will only be considered "prohibited" (with probability 0) if the empirical probability is 0 after Nmin or more occurences.}

\item{sensibility, alpha}{Alternatives to Nmin. If provided, \eqn{N_{min}=\log_{1-s}\alpha}where \eqn{s} is the sensibility. Both must be provided, see details for more.}

\item{cleantree}{If set to TRUE (default), non-leaf nodes' probability, dom, and context are emptied.}

\item{silent}{If TRUE, no progress messages are printed.}

\item{treeAsDataFrame}{should the tree output be the result of ToDataFrameTree()? If false, returns a data.tree object. Defaults to TRUE}
}
\value{
Skeleton of the VLMC adjusted to the sample provided, in two objects:

A \strong{Tree}, either in DataFrame format or tree object, with the columns/attributes:
\itemize{
\item \emph{context} context.
\item \emph{p} \eqn{p=P[X=1|X^{-1}_{-k}=w]}
\item \emph{n} number of occurences in the sample
\item \emph{dom} internal parameter to identify prohibited transitions. -1 if \eqn{p=0}, 1 if \eqn{p=1}, 0 if \eqn{p\in(0,1)}
}

A \strong{Skeleton Matrix}, determining which transitions have a postive probability. See \link{skel_matrix} for more details.
}
\description{
Generate skeleton tree
}
\details{
Since it is impossible to determine empirically that a probability is 0, a cutoff needs to be specified. Nmin is a simple but arbitrary way of doing it.
Sensibility and alpha: Probabilities greater or equal to \emph{sensibility} will be detected, at a \emph{alpha} confidence level.
}
