% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genskel.R
\name{genskel}
\alias{genskel}
\title{Tree generation}
\usage{
genskel(node, sample_vec, sample_txt, Nmin, prob = F)
}
\arguments{
\item{node}{Node to generate from.}

\item{sample_vec, sample_txt}{Different formats for the sample.}

\item{Nmin}{Cutoff for node generation, passes onto \link{genskel}. See \link{generate_skeleton.R} for more details.}

\item{prob}{Logical, should suffix probabilities be computed? Defaults to false.}
}
\description{
Internal recursive function to generate two nodes with context \strong{0w} and \strong{1w} from a node with context \strong{w}, as long as
\eqn{n > N_{min}} and the empirical probability of 1 is not 0 or 1.

Should not be used, only made for \link{startskel}.
}
\keyword{internal}
