% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.max.context.R
\name{get.max.context}
\alias{get.max.context}
\title{Get the biggest string that is a suffix of another}
\usage{
get.max.context(contexts, string)
}
\arguments{
\item{contexts}{Candidates.}

\item{string}{Bigger string for which we want a suffix.}
}
\value{
Biggest context matching the end of the string, or NULL if none match.
}
\description{
Finds the biggest string among \strong{contexts} that matches exactly the end of \strong{string}.
}
