% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMaxContext.R
\name{getMaxContext}
\alias{getMaxContext}
\title{Get the biggest string that is a suffix of another}
\usage{
getMaxContext(contexts, string)
}
\arguments{
\item{contexts}{Candidates.}

\item{string}{Bigger string for which we want a suffix.}
}
\value{
Biggest context matching the end of the string, or NULL if none match.
}
\description{
Finds the biggest string among \emph{contexts} that matches exactly the end of \emph{string}
}
