% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_hook.R
\name{ghook_sk_boilerplate}
\alias{ghook_sk_boilerplate}
\alias{ghook_sk_dashboard}
\title{golem_hook}
\usage{
ghook_sk_boilerplate(path, package_name, ...)

ghook_sk_dashboard(path, package_name, ...)
}
\arguments{
\item{path}{Name of the folder to create the package in.
This will also be used as the package name.}

\item{package_name}{Package name to use. By default, {golem} uses
\code{basename(path)}. If \code{path == '.'} & \code{package_name} is
not explicitly set, then \code{basename(getwd())} will be used.}

\item{...}{Arguments passed from \code{create_golem()}, unused in the default
function.}
}
\value{
used for side effects
}
\description{
golem_hook
}
\examples{
if (FALSE){
  golem::create_golem(
    "skboilerplate",
    project_hook = ghook_boilerplate
    )
  golem::create_golem(
    "skdashboard",
    project_hook = skeleton::ghook_dashboard
    )
}
}
