% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.check.R
\name{interval.check}
\alias{interval.check}
\title{Open and closed interval checking}
\usage{
interval.check(x, a, b, inc.L = F, inc.U = F)
}
\arguments{
\item{x}{x}

\item{a, b}{Lower and Upper bounds of interval}

\item{inc.L, inc.U}{Booleans, for having closed lower and upper bounds:
\itemize{
\item \strong{inc.L} = TRUE \eqn{\Rightarrow \;[a,}
\item \strong{inc.U} = TRUE \eqn{\Rightarrow \;\;,b]}
}}
}
\value{
checks if \emph{x} is in the constructed interval.
}
\description{
Open and closed interval checking
}
\examples{
library(skeleton)
interval_check(2,2,5)
interval_check(3,3,6,inc.L=T)

}
