% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_check.R
\name{interval_check}
\alias{interval_check}
\title{Interval checking}
\usage{
interval_check(x, a, b, inc.L = F, inc.U = F)
}
\arguments{
\item{x}{x}

\item{a, b}{Lower and Upper bounds of interval}

\item{inc.L, inc.U}{Booleans, for having closed lower and upper bounds:
\itemize{
\item \strong{inc.L} = TRUE \eqn{\Rightarrow \;[a,}
\item \strong{inc.U} = TRUE \eqn{\Rightarrow \;\;,b]}
}}
}
\value{
checks if \emph{x} is in the constructed interval.
}
\description{
Interval checking
}
