% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/killchildren.R
\name{killchildren}
\alias{killchildren}
\title{Skeleton Prune}
\usage{
killchildren(nodo)
}
\arguments{
\item{nodo}{Node whose children will be cut from the tree.}
}
\description{
Internal function to remove the children of a node. Doesn't return anything and shouldn't be used.
}
\keyword{internal}
