% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_str.R
\name{match_str}
\alias{match_str}
\title{Finding positions after a pattern in a string.}
\usage{
match_str(str, p)
}
\arguments{
\item{str}{String}

\item{p}{Pattern. An empty pattern returns 1:nchar(str).}
}
\value{
Vector of integer positions. If the pattern ends on the last character of the string, it won't be counted.
}
\description{
Finding positions after a pattern in a string.
}
\examples{
string = 'banana'
match_str(string, 'b')
match_str(string, 'an')
match_str(string, 'ana')
match_str(string, 'nana')
match_str(string, '')


}
