% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.to.trans.R
\name{matrix.to.trans}
\alias{matrix.to.trans}
\title{Converting a matrix of transitions to a list}
\usage{
matrix.to.trans(M, A, sep)
}
\arguments{
\item{M}{Matrix of transitions, with sequences of elements of \emph{A} of length k (order of the skeleton) for its rownames and colnames, where \deqn{M[w,w^{-1}_{-k+1}u]=I\{P[u|w]>0\}}}

\item{A}{Alphabet of the VLMC.}

\item{sep}{Character separator used in the generation of the skeleton.}
}
\value{
A list of named transition vectors.
}
\description{
Does the opposite of \link{trans.to.matrix}.
}
