% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sculptskeleton.R
\name{sculptskeleton}
\alias{sculptskeleton}
\title{Skeleton pruning}
\usage{
sculptskeleton(root, Nmin, copy = F, declare = F)
}
\arguments{
\item{root}{Tree, generated from \link{startskel}.}

\item{Nmin}{A transition will only be considered "prohibited" (with probability 0) if the empirical probability is 0 after Nmin or more occurences.}

\item{copy}{Set to TRUE to preserve the original tree and return a pruned copy.}

\item{declare}{If TRUE prints how many nodes are cut by level of the tree.}
}
\value{
If \strong{copy = TRUE}, returns the pruned skeleton. Otherwise, alters the original object with no output.
}
\description{
Prunes a probability tree generated by \link{startskel}. Made for \link{skeleton}.
}
\details{
For a node w, whose children are leaves, removes the children if:
\itemize{
\item both don't have enough occurrences to be significant.
\item both have the same prohibited transitions.
}

Starts from the second to last level of the tree.
}
