% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/should.you.cut.R
\name{should.you.cut}
\alias{should.you.cut}
\title{Wrapper for pruning}
\usage{
should.you.cut(parent, Nmin)
}
\arguments{
\item{parent}{Node to be evaluated.}

\item{Nmin}{Cutoff, see \link{skeleton} for details.}
}
\description{
Internal wrapper, should not be used.
}
\keyword{internal}
