% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shouldyoucut.R
\name{shouldyoucut}
\alias{shouldyoucut}
\title{Wrapper for pruning}
\usage{
shouldyoucut(parent, Nmin)
}
\arguments{
\item{parent}{Node to be evaluated.}

\item{Nmin}{Cutoff, see \link{generate_skeleton} for details.}
}
\description{
Internal wrapper, should not be used.
}
\keyword{internal}
