% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_buttons.R
\name{sk_button}
\alias{sk_button}
\alias{sk_button_primary}
\title{Buttons}
\usage{
sk_button(inputId, label, ...)

sk_button_primary(inputId, label, ...)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link--usually a text label, but
you could also use any other HTML, like an image.}

\item{...}{Named attributes to be applied to the button or link.}
}
\value{
an actionButton
}
\description{
Standard & primary buttons
}
\examples{
if (FALSE) {
  ui <- sk_page(
    h1("Hello world"),
    sk_row(
      sk_col(6, "Hello"),
      sk_col(6, "World"),
      sk_button("button", "Button"),
      sk_button_primary("button_primary", "Button primary")
    )
  )

  server <- function(input, output) {
    # Server code goes here
    observeEvent(input$button, {
      print("Hello standard button")
    })
    observeEvent(input$button_primary, {
      print("Hello button_primary")
    })
  }
  shinyApp(ui, server)
}
}
