% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_inputs.R
\name{sk_select_input}
\alias{sk_select_input}
\alias{sk_text_input}
\alias{sk_area_input}
\title{Inputs}
\usage{
sk_select_input(...)

sk_text_input(...)

sk_area_input(...)
}
\arguments{
\item{...}{arguments passed to the `selectInput` function}
}
\value{
an input tag (selectInput, textInput, textAreaInput)
}
\description{
Skeleton inputs
}
\examples{
if (FALSE) {
  ui <- sk_page(
    h1("Hello world"),
    sk_row(
      sk_col(
        6,
        sk_text_input(
          "email",
          "Your Email"
        )
      ),
      sk_col(
        6,
        sk_select_input(
          "reason",
          "Reason for contacting"
        )
      ),
      sk_col(
        12,
        sk_area_input(
          "message",
          "Message",
          placeholder = "hi Dave"
        )
      ),
      sk_button("send", "Send")
    )
  )

  server <- function(input, output) {
    # Server code goes here
    observeEvent(input$button, {
      print("Hello standard button")
    })
    observeEvent(input$button_primary, {
      print("Hello button_primary")
    })
  }
  shinyApp(ui, server)
}
}
