% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_nav.R
\name{sk_nav}
\alias{sk_nav}
\alias{sk_nav_item}
\title{navbar}
\usage{
sk_nav(...)

sk_nav_item(id, title, ui)
}
\arguments{
\item{...}{a list of navbar items}

\item{id}{the id of the navbar item}

\item{title}{the title of the navbar item}

\item{ui}{the ui of the navbar item}
}
\value{
a navbar
}
\description{
navbar
}
\examples{
if (FALSE){
  ui <- sk_page(
    sk_row(
      h2("A dead simple, responsive boilerplate."),
    ),
    sk_nav(
      sk_nav_item(
        id = "one",
        title = "ONE",
        ui = sk_row(
          sk_col(
            6,
            "Hello plot1"
          ),
          sk_col(
            6,
            plotOutput("plot1")
          )
        )
      ),
      sk_nav_item(
        id = "two",
        title = "TWO",
        ui = sk_row(
          sk_col(
            6,
            plotOutput("plot2")
          ),
          sk_col(
            6,
            "Hello plot2"
          ),
        )
      ),
      sk_nav_item(
        id = "three",
        title = "THREE",
        ui = sk_row(
          sk_col(
            6,
            plotOutput("plot3")
          ),
          sk_col(
            6,
            "Hello plot3"
          ),
        )
      )
    )
  )
  shinyApp(ui, function(input, output) {
    output$plot1 <- renderPlot({
      plot(mtcars)
    })
    output$plot2 <- renderPlot({
      plot(airquality)
    })
    output$plot3 <- renderPlot({
      plot(pressure)
    })
  })
}
}
