% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_page.R
\name{sk_page}
\alias{sk_page}
\alias{sk_row}
\alias{sk_col}
\alias{sk_col_one_third}
\alias{sk_col_one_half}
\alias{sk_col_two_third}
\alias{sk_header}
\title{Skeleton Page}
\usage{
sk_page(...)

sk_row(..., id = NULL)

sk_col(width, ..., id = NULL)

sk_col_one_third(..., id = NULL)

sk_col_one_half(..., id = NULL)

sk_col_two_third(..., id = NULL)

sk_header(title, ...)
}
\arguments{
\item{...}{The content of the element}

\item{id}{The id of the element}

\item{width}{The width of the element}

\item{title}{The title of the page}
}
\value{
A skeleton html element
}
\description{
Skeleton Page
}
\examples{
if (FALSE) {
  ui <- sk_page(
    tags$style(
      ".columns, .column {
        background: #EEE;
        text-align: center;
        border-radius: 4px;
        margin-bottom: 1.5rem;
      }"
    ),
    sk_row(
      h2("A dead simple, responsive boilerplate."),
    ),
    sk_row(
      tags$p(
        "The grid is a 12-column fluid grid with a max width of 960px,",
        "that shrinks with the browser/device at smaller sizes.",
        "The max width can be changed with one line of CSS and",
        "all columns will resize accordingly.",
        "The syntax is simple and it makes coding",
        "responsive much easier. Go ahead, resize the browser."
      )
    ),
    sk_row(
      sk_col(1, "ONE"),
      sk_col(11, "ELEVEN")
    ),
    sk_row(
      sk_col(2, "TWO"),
      sk_col(10, "TEN"),
    ),
    sk_row(
      sk_col(3, "THREE"),
      sk_col(9, "NINE")
    ),
    sk_row(
      sk_col(4, "FOUR"),
      sk_col(8, "EIGHT")
    ),
    sk_row(
      sk_col(5, "FIVE"),
      sk_col(7, "SEVEN")
    ),
    sk_row(
      sk_col(6, "SIX"),
      sk_col(6, "SIX")
    )
  )
  shinyApp(ui, function(input, output) {
    # Server code goes here
  })
}

}
