% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skel_matrix.R
\name{skel_matrix}
\alias{skel_matrix}
\title{Transition Skeleton Matrix}
\usage{
skel_matrix(tree)
}
\arguments{
\item{tree}{tree resulting from \link{sculptskeleton}. Must have context and p attributes if done manually.}
}
\value{
A \eqn{2^k\times 2^k} matrix \strong{M}, where \emph{k} is the maximum length of the \strong{tree}'s contexts.
}
\description{
Generates the transition skeleton matrix from a pruned skeleton. Meant for \link{generate_skeleton}.
}
\details{
Shorter contexts are expanded by pasting all possible pasts to their beginning. \\n
For example, if context w has k-2 symbols, \emph{00w, 10w, 01w,} and \emph{11w} will all be generated, with the same probability as w.

For a sequence \emph{w} of length \emph{k} \eqn{w=X^{-1}_{-k}}, its possible transitions are \eqn{w\_0=X^{-1}_{k+1}0, w\_1=X^{-1}_{k+1}1}

\eqn{M[w,w\_u]=1} if the transition \eqn{w, w\_u} has probability greater than 0, for \eqn{u\in {0,1}}
}
