\name{skeleton-package}
\alias{skeleton-package}
\alias{skeleton}
\docType{package}
\title{
\packageTitle{skeleton}
}
\description{
\packageDescription{skeleton}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{skeleton}
\packageIndices{skeleton}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{skeleton}

Maintainer: \packageMaintainer{skeleton}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
