% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startskel.R
\name{startskel}
\alias{startskel}
\title{Initializes the tree}
\usage{
startskel(sample_txt, Nmin, prob = F)
}
\arguments{
\item{sample_txt}{Data for the tree, in the format of a single string}

\item{Nmin}{Cutoff for node generation, passes onto \link{genskel}. See \link{generate_skeleton.R} for more details.}

\item{prob}{Logical, should suffix probabilities be computed? Defaults to false.}
}
\value{
Uncut suffix tree generated from the data as a \emph{data.tree} object.
}
\description{
Function to initialize a probability tree. Is intended for \link{generate_skeleton}, but can be used on its own to check the data.
Does not prune the tree at all.
}
