% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_to_vec.R
\name{string_to_vec}
\alias{string_to_vec}
\title{Converting integer string to a numeric vector.}
\usage{
string_to_vec(string)
}
\arguments{
\item{string}{String comprised of positive integers, including 0. Other values won't work.}
}
\value{
Vector of integers.
}
\description{
Efficient function made for the skeleton package, might be useful for other cases.
}
\examples{
string_to_vec('1001')


}
