% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.to.matrix.R
\name{trans.to.matrix}
\alias{trans.to.matrix}
\title{Converting a list of transitions to a matix}
\usage{
trans.to.matrix(transitions, A, sep)
}
\arguments{
\item{transitions}{List of size \eqn{|A|^k}, containing named binary vectors of size \eqn{|A|}.}

\item{A}{Alphabet of the VLMC.}

\item{sep}{Character separator used in the generation of the skeleton.}
}
\value{
A transition matrix. See \link{skeleton} for more details.
}
\description{
Given a sorted alphabet \eqn{A = \{a_1,a_2,...,a_{|A|}\}} and a list of named vectors \eqn{v^{(w)}},
where \eqn{w} is a context and
\deqn{v^{(w)}_i = I\{P[a_i|w]>0\}}
will return the matrix \eqn{M}, where
\deqn{M[w,w^{-1}_{-k+1}a_i]=v^{(w)}_i}
}
