% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transToMatrix.R
\name{transToMatrix}
\alias{transToMatrix}
\title{Converting a list of transitions to a matix}
\usage{
transToMatrix(transitions, A, sep)
}
\arguments{
\item{transitions}{List of transitions, vectors \eqn{v} of length \eqn{|A|} and name \eqn{w} where \deqn{v_i = I[P[a_i|w]>0],\;\forall a_i\in A}}

\item{A}{Alphabet of the VLMC.}

\item{sep}{Character separator used in the generation of the skeleton.}
}
\value{
A transition matrix. See \link{skel_matrix} for more details.
}
\description{
Converting a list of transitions to a matix
}
