% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{use_skeleton}
\alias{use_skeleton}
\title{Create a new project}
\usage{
use_skeleton()
}
\description{
From a newly created project, this function sets up a skeleton for a data
analysis project in R. The skeleton includes directories for data, code, and
results; a readme; a basic drake plan; a lab notebook using bookdown; and an
R Markdown document for tables and figures.
}
