% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_to_string.R
\name{vec_to_string}
\alias{vec_to_string}
\title{Converting integer numeric vector to binary string.}
\usage{
vec_to_string(vector)
}
\arguments{
\item{vector}{Vector of positive integers, including 0. Other values won't work.}
}
\value{
String of numeric characters.
}
\description{
Efficient function made for the skeleton package, might be useful for other cases.
}
\examples{
vec_to_string(c(0,1,1,0))


}
