% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_skeleton.R
\name{generate_skeleton}
\alias{generate_skeleton}
\title{Generate skeleton tree}
\usage{
generate_skeleton(
  sample,
  Nmin = 59,
  sensibility,
  alpha,
  output = "both",
  cleantree = T
)
}
\arguments{
\item{sample}{Data. Either a string or a numeric vector. Should be longer than Nmin and only contain 0 and 1 symbols.}

\item{Nmin}{A transition will only be considered "prohibited" (with probability 0) if the empirical probability is 0 after Nmin or more occurences.}

\item{sensibility, alpha}{Alternatives to Nmin. If provided, \eqn{N_{min}=\log_{1-s}\alpha}where \eqn{s} is the sensibility. Both must be provided, see details for more.}

\item{output}{"tree", "matrix", "both" or the first letter of each. Defaults to both.}

\item{cleantree}{If set to TRUE (default), non-leaf nodes' probability, dom, and context are emptied.}
}
\value{
Skeleton of the VLMC adjusted to the sample provided, either as the suffix tree (data.tree object) or the transition matrix.
}
\description{
Generate skeleton tree
}
\details{
Since it is impossible to determine empirically that a probability is 0, a cutoff needs to be specified. Nmin is a simple but arbitrary way of doing it.
Sensibility and alpha: Probabilities greater or equal to \emph{sensibility} will be detected, at a \emph{alpha} confidence level.
}
