% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genskel.R
\name{genskel}
\alias{genskel}
\title{Recursive Tree Generation}
\usage{
genskel(parent, level, sample, alphabet, Nmin, sep = "-", contextsep = sep)
}
\arguments{
\item{parent}{Node to generate from.}

\item{level}{Level of the node in the tree, number of symbols in its context.}

\item{sample}{Data in the form of an atomic vector. Can only contain symbols present in \emph{alphabet}, and needs to be longer than \strong{Nmin}.}

\item{alphabet}{Set \eqn{A} of possible symbols for the VLMC. Defaults to sort(unique(sample)).}

\item{Nmin}{A transition will only be considered "prohibited" (with probability 0) if the empirical probability is 0 after Nmin or more occurences.}

\item{sep}{Symbol separator for the node names.}

\item{contextsep}{Symbol separator for the node contexts.}
}
\description{
Internal recursive function to generate nodes with context \emph{uw} \eqn{\forall u \in A} from a node with context \emph{w}, as long as
\eqn{n > N_{min}} and two or more transitions are possible.

Should not be used, only made for \link{startskel}.
}
\keyword{internal}
