% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_check.R
\name{interval_check}
\alias{interval_check}
\title{Interval checking}
\usage{
interval_check(x, a, b, inc.L = F, inc.U = F)
}
\arguments{
\item{x}{x}

\item{a, b}{lower and upper bounds of interval}

\item{inc.L, inc.U}{Logical, should the interval be closed?}
}
\value{
checks if \emph{x} is in \eqn{(a,b),`[`a,b),(a,b`]`,`[a,b]`}, depending on the choices for \emph{inc.L} and \emph{inc.U}
}
\description{
Interval checking
}
