% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irreductible.R
\name{irreductible}
\alias{irreductible}
\title{Irreductibility of the Skeleton Matrix}
\usage{
irreductible(M, ReturnMatrix = FALSE)
}
\arguments{
\item{M}{A binary square matrix generated by \link{generate_skeleton} or \link{skel_matrix}.}

\item{ReturnMatrix}{Do you want Q to be returned? Defaults to FALSE. See Details.}
}
\value{
Either a logical result for irreductibility or a list containing the result and Q.
}
\description{
Tests the irreductibility of the Skeleton Transition Matrix
}
\details{
A skeleton of order \eqn{k} will result in a transition matrix \eqn{M_{2^k\times 2^k}}.

M is irreductible if
\deqn{\forall i,j \in [1,2^k], \exists n \in ℕ | M^n[i,j]>0}
In practice this can be tested with \eqn{n = 1,...,2^k-1}.\deqn{Q[i,j] = \mathcal{I}\{(\sum_{l=1}^{2^k}M^l)[i,j]>0\}}
}
