% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_str.R
\name{match_str}
\alias{match_str}
\title{Finding positions after a pattern in a string.}
\usage{
match_str(str, p)
}
\arguments{
\item{str}{string}

\item{p}{pattern}
}
\value{
Integer positions: if \code{str[3:8]} = p, returns 9.
}
\description{
Finding positions after a pattern in a string.
}
