% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shouldyoucut.R
\name{shouldyoucut}
\alias{shouldyoucut}
\title{Wrapper for pruning}
\usage{
shouldyoucut(nodo, par, Nmin)
}
\arguments{
\item{nodo}{node}

\item{par}{name of the parameter to base cutting}

\item{Nmin}{cutoff, see \link{generate_skeleton} for details}
}
\description{
Internal wrapper, should not be used.
}
\keyword{internal}
