% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startskel.R
\name{startskel}
\alias{startskel}
\title{Initializes the tree}
\usage{
startskel(sample, alphabet, Nmin, sep = "-", contextsep = sep)
}
\arguments{
\item{sample}{Data in the form of an atomic vector. Can only contain symbols present in \emph{alphabet}, and needs to be longer than \strong{Nmin}.}

\item{alphabet}{Set \eqn{A=\{a_1,a_2,...,a_{|A|}\}} of possible symbols for the VLMC. Defaults to sort(unique(sample)).}

\item{Nmin}{A transition will only be considered "prohibited" (with probability 0) if the empirical probability is 0 after Nmin or more occurences.}

\item{sep, contextsep}{Symbol separator for the node names. A separate symbol can be specified for the contexts, if wanted. Both symbols cannot be present in \eqn{A}.}
}
\value{
Uncut suffix tree generated from the data as a \emph{data.tree} object.
}
\description{
Function to initialize a probability tree. Is intended for \link{skeleton}, but can be used on its own to check the data.
Does not prune the tree at all. For manual usage, do \link{startskel} -> \link{sculptskeleton} -> \link{expand.transitions} -> \link{trans.to.matrix}.
}
\examples{
library(skeleton)
(t = startskel(c('0','1','0','0','0','1','1','0'), c('0','1'), Nmin = 1))
ToDataFrameTree(t, 'context', 'n', 'counts')


}
