% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_to_vec.R
\name{string_to_vec}
\alias{string_to_vec}
\title{Converting binary string to a numeric vector}
\usage{
string_to_vec(string)
}
\arguments{
\item{string}{}
}
\value{
'101' returns c(1,0,1)
}
\description{
Converting binary string to a numeric vector
}
