% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_to_string.R
\name{vec_to_string}
\alias{vec_to_string}
\title{Converting binary numeric vector to binary string}
\usage{
vec_to_string(vector)
}
\arguments{
\item{vector}{}
}
\value{
c(1,0,1) returns '101'
}
\description{
Converting binary numeric vector to binary string
}
